+++ config/horizon.php
@@ -339,6 +339,19 @@ return [
             ],
             'supervisor-5' => [
                 'connection' => 'redis',
+                'queue' => ['tests'],
+                'balance' => 'auto',
+                'autoScalingStrategy' => 'time',
+                'maxProcesses' => 3,
+                'maxTime' => 0,
+                'maxJobs' => 0,
+                'memory' => 128,
+                'tries' => 1,
+                'timeout' => 60,
+                'nice' => 0,
+            ],
+            'supervisor-6' => [
+                'connection' => 'redis',
                 'queue' => ['long-running'],
                 'balance' => 'auto',
                 'autoScalingStrategy' => 'time',
@@ -352,7 +365,7 @@ return [
                 'timeout' => 300,
                 'nice' => 0,
             ],
-            'supervisor-6' => [
+            'supervisor-7' => [
                 'connection' => 'redis',
                 'queue' => ['schedules'],
                 'balance' => 'auto',