+++ resources/js/Components/Global/Feedback/OrangeFeedbackBox.vue
@@ -4,30 +4,9 @@
          @mouseover="hover = true"
          @mouseleave="hover = false"
          @click="openModal"
-         ></div>
+    ></div>
     <div
-         class="h-custom w-3 bg-orange-500 z-100 text-white">
-
-
-<!--      <div class="fixed z-50 left-3 bottom-24 w-28 h-14 bg-orange-950 hover:bg-blue-950 bg-opacity-80">-->
-<!--        <div class="pl-2 uppercase font-semibold text-sm tracking-wider ">-->
-<!--          Feedback-->
-<!--        </div>-->
-<!--        <div class="pl-2 text-xs tracking-wider w-36 pr-6">-->
-<!--          Found a bug? Have a question?-->
-<!--        </div>-->
-<!--      </div>-->
-
-<!--      <div class="group fixed z-50 left-3 bottom-24 w-28 h-14 bg-orange-950 bg-opacity-0 hover:bg-opacity-80 bg-opacity-transition hover:bg-blue-950 flex flex-col justify-center hover:flex-row transition-all duration-300">-->
-<!--        <div class="pl-2 uppercase font-semibold text-sm tracking-wider rotate-90 group-hover:rotate-0 transition-transform">-->
-<!--          Feedback-->
-<!--        </div>-->
-<!--        <div class="pl-2 text-xs tracking-wider w-36 pr-6 opacity-0 group-hover:opacity-100 transition-opacity duration-300">-->
-<!--          Found a bug? Have a question?-->
-<!--        </div>-->
-<!--      </div>-->
-
-
+        class="h-custom w-3 bg-orange-500 z-100 text-white">
 
       <div :class="[
     'fixed z-50 left-3 bottom-36 w-32 h-custom flex flex-col justify-center transition-all duration-300 bg-opacity-transition',
@@ -41,7 +20,7 @@
       notificationStore.showOrangeFeedbackBox || hover ? 'rotate-0' : 'rotate-90',
     ]">
           <span
-                :class="[notificationStore.showOrangeFeedbackBox || hover ? '' : 'bg-orange-950 bg-opacity-80 px-2 pb-1']">Feedback</span>
+              :class="[notificationStore.showOrangeFeedbackBox || hover ? '' : 'bg-orange-950 bg-opacity-80 px-2 pb-1']">Need Help?</span>
         </div>
         <div class="pl-2 text-xs tracking-wider w-36 pr-6 transition-opacity duration-300"
              :class="[notificationStore.showOrangeFeedbackBox || hover ? 'opacity-100' : 'opacity-0']"
@@ -55,6 +34,7 @@
 
   </div>
 
+  <HandAnimation direction="left" />
 
   <!-- Open the modal using ID.showModal() method -->
   <dialog id="feedbackConfirmationMessage" class="modal">
@@ -76,7 +56,7 @@
   <dialog id="orangeFeedbackModal" class="modal text-black">
 
     <div>
-      <div class="modal-box">
+      <div class="modal-box bg-gray-200 text-black">
         <div class="flex flex-col text-center justify-center w-full">
           <div class="tracking-wider uppercase text-orange-500">
             Feedback Form
@@ -84,7 +64,8 @@
         </div>
 
         <h3 class="font-bold text-lg">Hello!</h3>
-        <p class="py-4">Thank you for being one of the first to test out notTV. Use this form to send Cathy and Travis a message. <br /><br />If you have any problems with the website
+        <p class="py-4">Thank you for being one of the first to test out notTV. Use this form to send Cathy and Travis a
+          message. <br/><br/>If you have any problems with the website
           please include a detailed description.</p>
         <div>
           <textarea class="rounded-lg w-full" placeholder="message..." tabindex="0" v-model="form.message"></textarea>
@@ -118,6 +99,7 @@ import { useForm } from '@inertiajs/vue3'
 import { usePage } from '@inertiajs/vue3'
 import { computed, ref } from 'vue'
 import { router } from '@inertiajs/vue3'
+import HandAnimation from '@/Components/Global/Animations/HandAnimation.vue'
 
 const appSettingStore = useAppSettingStore()
 const notificationStore = useNotificationStore()
@@ -125,8 +107,8 @@ const notificationStore = useNotificationStore()
 const page = usePage()
 const user = computed(() => page.props.auth.user)
 
-const screenshot = ref(null);
-const hover = ref(false);
+const screenshot = ref(null)
+const hover = ref(false)
 
 let form = useForm({
   message: '',
@@ -138,32 +120,20 @@ let form = useForm({
 })
 
 function handleFileChange(event) {
-  screenshot.value = event.target.files[0];
+  screenshot.value = event.target.files[0]
 }
 
-// const submit = () => {
-//   let formData = new FormData();
-//   // Append other form data
-//   formData.append('screenshot', form.value.screenshot);
-//   router.post('/user/feedback', form)
-//   form.message = ''
-//   closeModal()
-//   document.getElementById('feedbackConfirmationMessage').showModal()
-// }
-
-
-
 const submit = () => {
-  const formData = new FormData();
+  const formData = new FormData()
   // Append other form data
   // Append text fields from the form
   Object.keys(form.data()).forEach(key => {
-    formData.append(key, form.data()[key]);
-  });
+    formData.append(key, form.data()[key])
+  })
 
   // Append the file if it exists
   if (screenshot) {
-    formData.append('screenshot', screenshot);
+    formData.append('screenshot', screenshot)
   }
 
   // Use Inertia's post method directly with formData
@@ -177,27 +147,19 @@ const submit = () => {
 
     // Optionally, handle onSuccess, onError, etc.
     onSuccess: () => {
-      console.log('Form submitted successfully');
+      console.log('Form submitted successfully')
       form.message = ''
       closeModal()
       document.getElementById('feedbackConfirmationMessage').showModal()
     },
     onError: () => {
-      console.log('Error submitting the form');
+      console.log('Error submitting the form')
       form.message = ''
       closeModal()
       document.getElementById('feedbackConfirmationMessage').showModal()
-    }
-  });
+    },
+  })
 
-
-
-
-  // formData.append('screenshot', form.value.screenshot);
-  // router.post('/user/feedback', form)
-  // form.message = ''
-  // closeModal()
-  // document.getElementById('feedbackConfirmationMessage').showModal()
 }
 
 const openModal = () => {
@@ -217,6 +179,7 @@ const closeModal = () => {
 .h-custom {
   height: 6rem; /* or whatever value you choose */
 }
+
 /* Custom CSS for rotating the feedback text */
 .rotate-90 {
   transform: rotate(-90deg);